+++ packages/commercetools/composables/src/useCart/index.ts
@@ -1,78 +1,69 @@
-import {
-  addToCart as apiAddToCart,
-  removeFromCart as apiRemoveFromCart,
-  updateCartQuantity as apiUpdateCartQuantity,
-  applyCartCoupon as apiApplyCartCoupon,
-  removeCartCoupon as apiRemoveCartCoupon,
-  getSettings,
-  isTokenUserSession,
-  CustomQueryFn
-} from '@vue-storefront/commercetools-api';
+import { CustomQueryFn } from '@vue-storefront/commercetools-api';
 import { ProductVariant, Cart, LineItem } from './../types/GraphQL';
 import loadCurrentCart from './currentCart';
-import { AgnosticCoupon, useCartFactory, UseCartFactoryParams } from '@vue-storefront/core';
+import { AgnosticCoupon, useCartFactory, UseCartFactoryParams, Context } from '@vue-storefront/core';
 
 const getBasketItemByProduct = ({ currentCart, product }) => {
   return currentCart.lineItems.find((item) => item.productId === product._id);
 };
 
 /** returns current cart or creates new one **/
-const getCurrentCart = async (currentCart) => {
+const getCurrentCart = async (context: Context, currentCart) => {
   if (!currentCart) {
-    return loadCurrentCart();
+    return loadCurrentCart(context);
   }
 
   return currentCart;
 };
 
 const params: UseCartFactoryParams<Cart, LineItem, ProductVariant, AgnosticCoupon> = {
-  loadCart: async (CustomQueryFn?: any) => {
-    const settings = getSettings();
+  loadCart: async (context: Context, CustomQueryFn?: any) => {
+    const { $ct } = context;
 
-    if (!isTokenUserSession(settings.currentToken)) {
+    if ($ct.api.isGuest()) {
       return null;
     }
 
-    return await loadCurrentCart(CustomQueryFn);
+    return await loadCurrentCart(context, CustomQueryFn);
   },
-  addToCart: async ({ currentCart, product, quantity }, customQuery?: CustomQueryFn) => {
-    const loadedCart = await getCurrentCart(currentCart);
+  addToCart: async (context: Context, { currentCart, product, quantity }, customQuery?: CustomQueryFn) => {
+    const loadedCart = await getCurrentCart(context, currentCart);
 
-    const { data } = await apiAddToCart(loadedCart, product, quantity, customQuery);
+    const { data } = await context.$ct.api.addToCart(loadedCart, product, quantity, customQuery);
     return data.cart;
   },
-  removeFromCart: async ({ currentCart, product }, customQuery?: CustomQueryFn) => {
-    const loadedCart = await getCurrentCart(currentCart);
+  removeFromCart: async (context: Context, { currentCart, product }, customQuery?: CustomQueryFn) => {
+    const loadedCart = await getCurrentCart(context, currentCart);
 
-    const { data } = await apiRemoveFromCart(loadedCart, product, customQuery);
+    const { data } = await context.$ct.api.removeFromCart(loadedCart, product, customQuery);
     return data.cart;
   },
-  updateQuantity: async ({ currentCart, product, quantity }, customQuery?: CustomQueryFn) => {
-    const loadedCart = await getCurrentCart(currentCart);
+  updateQuantity: async (context: Context, { currentCart, product, quantity }, customQuery?: CustomQueryFn) => {
+    const loadedCart = await getCurrentCart(context, currentCart);
 
-    const { data } = await apiUpdateCartQuantity(loadedCart, { ...product, quantity }, customQuery);
+    const { data } = await context.$ct.api.updateCartQuantity(loadedCart, { ...product, quantity }, customQuery);
     return data.cart;
   },
-  clearCart: async ({ currentCart }) => {
+  clearCart: async (context: Context, { currentCart }) => {
     return currentCart;
   },
-  applyCoupon: async ({ currentCart, couponCode }, customQuery?: CustomQueryFn) => {
-    const loadedCart = await getCurrentCart(currentCart);
+  applyCoupon: async (context: Context, { currentCart, couponCode }, customQuery?: CustomQueryFn) => {
+    const loadedCart = await getCurrentCart(context, currentCart);
 
-    const { data } = await apiApplyCartCoupon(loadedCart, couponCode, customQuery);
+    const { data } = await context.$ct.api.applyCartCoupon(loadedCart, couponCode, customQuery);
     return { updatedCart: data.cart, updatedCoupon: couponCode };
   },
-  removeCoupon: async ({ currentCart, coupon }, customQuery?: CustomQueryFn) => {
-    const loadedCart = await getCurrentCart(currentCart);
+  removeCoupon: async (context: Context, { currentCart, coupon }, customQuery?: CustomQueryFn) => {
+    const loadedCart = await getCurrentCart(context, currentCart);
 
-    const { data } = await apiRemoveCartCoupon(loadedCart, { id: coupon.id, typeId: 'discount-code' }, customQuery);
+    const { data } = await context.$ct.api.removeCartCoupon(loadedCart, { id: coupon.id, typeId: 'discount-code' }, customQuery);
     return { updatedCart: data.cart };
   },
-  isOnCart: ({ currentCart, product }) => {
+  isOnCart: (context: Context, { currentCart, product }) => {
     return Boolean(currentCart && getBasketItemByProduct({ currentCart, product }));
   }
 };
 
-const { useCart, setCart } = useCartFactory<Cart, LineItem, ProductVariant, AgnosticCoupon>(params);
+const { useCart } = useCartFactory<Cart, LineItem, ProductVariant, AgnosticCoupon>(params);
 
-export { useCart, setCart };
+export { useCart };