+++ packages/commercetools/composables/src/useReview/index.ts
@@ -1,15 +1,15 @@
 /* istanbul ignore file */
 
-import { useReviewFactory, UseReview, UseReviewFactoryParams } from '@vue-storefront/core';
+import { useReviewFactory, UseReview, UseReviewFactoryParams, Context } from '@vue-storefront/core';
 
 const params: UseReviewFactoryParams<any, any, any> = {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  searchReviews: async (params) => {
+  searchReviews: async (context: Context, params) => {
     console.log('Mocked: searchReviews');
     return {};
   },
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  addReview: async (params) => {
+  addReview: async (context: Context, params) => {
     console.log('Mocked: addReview');
     return {};
   }