+++ packages/core/nuxt-module/lib/modules/storefront-ui.js
@@ -1,6 +1,6 @@
 const merge = require('../helpers/merge');
-const resolveDependency = require('../helpers/resolveDependency');
 
+// TODO: Create a separate nuxt module for storefront ui
 function loadStorefrontRawSources (options) {
   const rawSources = [
     '@storefront-ui/vue',
@@ -13,19 +13,6 @@ function loadStorefrontRawSources (options) {
   });
 }
 
-function loadStorefrontStylesAndVariables () {
-  this.options.styleResources = {
-    scss: [
-      resolveDependency('@storefront-ui/shared/styles/_helpers.scss')
-    ]
-  };
-
-  this.options.css.push(resolveDependency('@storefront-ui/vue/styles.scss'));
-
-  this.addModule('@nuxtjs/style-resources');
-}
-
 module.exports = function VueStorefrontPerformanceModule (options) {
   loadStorefrontRawSources.call(this, options);
-  loadStorefrontStylesAndVariables.call(this);
 };